Move zone related function into zone file
