+++ configdns-v1/service.go
@@ -1,7 +1,6 @@
 package dns
 
 import (
-	"github.com/akamai/AkamaiOPEN-edgegrid-golang/client-v1"
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/edgegrid"
 )
 
@@ -9,35 +8,3 @@ var (
 	// Config contains the Akamai OPEN Edgegrid API credentials
 	Config edgegrid.Config
 )
-
-// GetZone retrieves a DNS Zone for a given hostname
-func GetZone(hostname string) (*Zone, error) {
-	zone := NewZone(hostname)
-	req, err := client.NewRequest(
-		Config,
-		"GET",
-		"/config-dns/v1/zones/"+hostname,
-		nil,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := client.Do(Config, req)
-	if err != nil {
-		return nil, err
-	}
-
-	if client.IsError(res) && res.StatusCode != 404 {
-		return nil, client.NewAPIError(res)
-	} else if res.StatusCode == 404 {
-		return nil, &ZoneError{zoneName: hostname}
-	} else {
-		err = client.BodyJSON(res, &zone)
-		if err != nil {
-			return nil, err
-		}
-
-		return zone, nil
-	}
-}
